Refactor CollectionsPage to use Suspense for loading state and create LoadingState component for better loading experience
